chore: resolve cyclic dependency
